requirements.txt
aiohttp>=3.10.10
asyncio>=3.4.3
beautifulsoup4>=4.12.0
cachetools>=5.5.0
ccxt==4.1.13
cryptocompare>=0.7.6
feedparser>=6.0.10
nest-asyncio
nltk>=1.6.0
numpy==1.24.3
openai>=1.54.3
pandas==2.0.3
plotly==5.18.0
pycoingecko==3.1.0
python-dateutil
python-dotenv==1.0.0
python-telegram-bot>=2.9.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.3.2
seaborn==0.13.0
spacy==3.7.2
streamlit==1.24.0
ta
textblob>=0.18.0
tweepy>=4.14.0
vadersentiment>=3.3.2
websockets>=14.1
yfinance==0.2.31
TA-Lib>=0.4.28
praw>=7.7.1
